+++ railties/lib/rails/generators/test_unit/scaffold/scaffold_generator.rb
@@ -14,7 +14,7 @@ module TestUnit # :nodoc:
                          desc: "Generate API functional tests"
 
       class_option :system_tests, type: :string,
-                         desc: "Skip system test files"
+                         desc: "Generate system test files (set to 'true' to enable)"
 
       argument :attributes, type: :array, default: [], banner: "field:type field:type"
 
@@ -23,7 +23,9 @@ module TestUnit # :nodoc:
         template template_file,
                  File.join("test/controllers", controller_class_path, "#{controller_file_name}_controller_test.rb")
 
-        if !options.api? && options[:system_tests]
+        # Generate system tests if this isn't an API only app and the system
+        # tests option is true
+        if !options.api? && options[:system_tests] == "true"
           template "system_test.rb", File.join("test/system", class_path, "#{file_name.pluralize}_test.rb")
         end
       end