+++ railties/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Don't generate system tests by default.
+
+    Rails scaffold generator will no longer generate system tests by default. To enable this pass `--system-tests=true` or generate them with `bin/rails generate system_test name_of_test`.
+
+    *Eileen M. Uchitelle*
+
 *   Optionally skip bundler-audit.
 
     Skips adding the `bin/bundler-audit` & `config/bundler-audit.yml` if the gem is not installed when `bin/rails app:update` runs.